refactor: voice.rs phase 2 — extract remaining 13 renderers (UPI 050)
Closes the UPI 050 split. Phase 1 (PR #142) extracted doctor + status;
phase 2 extracts the remaining 13 per-command renderers from
voice/mod.rs into sibling sub-modules under src/voice/.

New sibling modules:
- voice/backup.rs       — render_backup_summary + render_pre_action
                          + transitions/skipped-block/assessment-table
- voice/plan.rs         — render_plan + render_empty_plan
                          + skip-group helpers
- voice/verify.rs       — render_verify + render_failures
- voice/history.rs      — render_history + render_subvolume_history
                          + render_events
- voice/init.rs         — render_init + format_init_status
- voice/calibrate.rs    — render_calibrate
- voice/sentinel.rs     — render_sentinel_status + format_tick_description
- voice/emergency.rs    — render_emergency + render_emergency_result
- voice/retention.rs    — render_retention_preview + breakdown helper
- voice/drives.rs       — render_drives_list + render_drives_adopt
                          + status/token formatters + absent-duration
- voice/get.rs          — render_get (daemon + interactive)
- voice/chooser.rs      — format_subvolume_chooser (+ its 2 tests)
- voice/status.rs       — extended with render_default_status
                          + render_first_time

Cross-renderer helpers stayed in voice/mod.rs as pub(super):
humanize_duration, exposure_label, format_status_table, color_* family,
pluralize, classify_verify_checks, append_suggestion + SuggestionContext,
format_history_table, truncate_str, skip_tag, aggregate_drive_info,
unmounted_drive_label, format_drive_age_label, status_severity.

skip_tag is a new pub(super) helper (canonical home: voice/mod.rs)
since both plan.rs and backup.rs render skip tags; previously private
to mod.rs as a plan-section local.

format_history_table and truncate_str were promoted from fn to
pub(super) in phase 2 — used by both history.rs and verify.rs.

voice/mod.rs went from 8008 lines (post-phase-1) to ~6000 lines.
Public surface preserved via re-exports: voice::render_backup_summary,
voice::render_plan, etc. — 23 caller sites in src/commands/ unchanged.

One test moved: backup_summary_suppresses_unchanged (which directly
exercised render_skipped_block) moved from voice/mod.rs::mod tests
to voice/backup.rs::mod tests, since render_skipped_block now lives
in backup.rs and is private to that sub-module.

cargo clippy --all-targets -- -D warnings + cargo test (1435 passed) +
cargo build --release all green. Voice Contract suite (44 tests) green
pre- and post-split — identical golden strings.
