+++ CLAUDE.md
@@ -76,7 +76,7 @@ All backup logic flows through: config -> plan -> execute. No exceptions.
 | `drives.rs` | Detect mounted drives, UUID fingerprinting, check space | Mount/unmount drives |
 | `pools.rs` | Detect BTRFS pools (source + destination), group subvolumes by pool UUID, read sysfs metadata utilization and statvfs free bytes | Know about retention, plans, drive lifecycle, or notification policy |
 | `output.rs` | Define structured output types | Render text (voice.rs does that) |
-| `voice/` | Render structured output as text (mythic voice). Per-command sub-modules under `voice/` (currently `doctor.rs`, `status.rs`; remaining renderers still in `voice/mod.rs` pending UPI 050 phase 2). Cross-renderer helpers (`humanize_duration`, `exposure_label`, `color_*`, `pluralize`, `classify_verify_checks`, `append_suggestion`) live in `voice/mod.rs` and are `pub(super)` so sub-modules can use them | Perform I/O or compute state |
+| `voice/` | Render structured output as text (mythic voice). Per-command sub-modules under `voice/` (`backup.rs`, `calibrate.rs`, `chooser.rs`, `doctor.rs`, `drives.rs`, `emergency.rs`, `get.rs`, `history.rs`, `init.rs`, `plan.rs`, `retention.rs`, `sentinel.rs`, `status.rs`, `verify.rs`) — the UPI 050 per-command split is now complete. Cross-renderer helpers (`humanize_duration`, `exposure_label`, `color_*`, `pluralize`, `classify_verify_checks`, `append_suggestion`/`SuggestionContext`, `format_history_table`, `truncate_str`, `skip_tag`, `aggregate_drive_info`, `unmounted_drive_label`, `format_drive_age_label`, `status_severity`) live in `voice/mod.rs` and are `pub(super)` so sub-modules can use them. Public surface preserved via re-exports — callers continue to use `voice::render_*` unchanged | Perform I/O or compute state |
 | `voice_events.rs` | Per-variant `EventPayload` renderer (columnar + NDJSON) | Perform I/O or query state |
 | `events.rs` | Pure: `Event`, `EventKind`, `EventPayload`, `Severity`, typed payload enums (UPI 036) | Perform I/O |
 | `lock.rs` | Shared advisory lock with metadata (PID, trigger source) | Decide whether to proceed (caller's job) |
