+++ src/voice/mod.rs
@@ -11,31 +11,50 @@ use std::fmt::Write;
 
 use colored::Colorize;
 
-use crate::output::{
-    AdoptAction, BackupSummary, CalibrateOutput, CalibrateResult,
-    DefaultStatusOutput,
-    DriveAdoptOutput, DriveStatus, DrivesListOutput, EmergencyOutput, EmergencyResult,
-    FailuresOutput, GetOutput, HistoryOutput, InitOutput, InitStatus, OutputMode, PlanOutput,
-    PreActionSummary, RecoveryWindow, RetentionPreviewOutput,
-    SentinelStatusOutput, SkipCategory, SkippedSubvolume, StatusAssessment,
-    SubvolumeHistoryOutput, TokenState, VerifyCheck, VerifyOutput, parse_duration_to_minutes,
-};
+use crate::output::{SkipCategory, StatusAssessment, VerifyCheck, VerifyOutput};
 // Test-only imports — moved-renderer types still referenced by parent tests
-// (status/doctor renderer extraction to UPI 050).
+// (renderer extractions per UPI 050 phases 1 + 2).
 #[cfg(test)]
 use crate::output::{
-    ChainHealth, DoctorCheck, DoctorCheckStatus, DoctorOutput, StatusOutput,
+    AdoptAction, BackupSummary, ChainHealth, DefaultStatusOutput, DoctorCheck, DoctorCheckStatus,
+    DoctorOutput, DriveAdoptOutput, DrivesListOutput, EmergencyOutput, EmergencyResult, OutputMode,
+    PlanOutput, PreActionSummary, RecoveryWindow, RetentionPreviewOutput, SentinelStatusOutput,
+    SkippedSubvolume, StatusOutput,
 };
-use crate::plan::format_duration_short;
+#[cfg(test)]
 use crate::types::{ByteSize, DriveRole};
 
 // ── Sub-modules (per-command renderers; UPI 050) ──────────────────────
 
+mod backup;
+mod calibrate;
+mod chooser;
 mod doctor;
+mod drives;
+mod emergency;
+mod get;
+mod history;
+mod init;
+mod plan;
+mod retention;
+mod sentinel;
 mod status;
+mod verify;
 
+pub use backup::{render_backup_summary, render_pre_action};
+pub use calibrate::render_calibrate;
+pub use chooser::format_subvolume_chooser;
 pub use doctor::render_doctor;
-pub use status::render_status;
+pub use drives::{render_drives_adopt, render_drives_list};
+pub use emergency::{render_emergency, render_emergency_result};
+pub use get::render_get;
+pub use history::{render_events, render_history, render_subvolume_history};
+pub use init::render_init;
+pub use plan::{render_empty_plan, render_plan};
+pub use retention::render_retention_preview;
+pub use sentinel::render_sentinel_status;
+pub use status::{render_default_status, render_first_time, render_status};
+pub use verify::{render_failures, render_verify};
 
 // ── Cross-renderer helpers ────────────────────────────────────────────
 
@@ -213,747 +232,13 @@ pub(super) fn color_result(result: &str) -> String {
     }
 }
 
-// ── Backup Summary ─────────────────────────────────────────────────────
 
-/// Render post-backup summary according to the given mode.
-#[must_use]
-pub fn render_backup_summary(data: &BackupSummary, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_backup_interactive(data),
-        OutputMode::Daemon => render_backup_daemon(data),
-    }
-}
-
-fn render_backup_daemon(data: &BackupSummary) -> String {
-    serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-}
-
-fn render_backup_interactive(data: &BackupSummary) -> String {
-    let mut out = String::new();
-
-    // ── Header ───────────────────────────────────────────────────────
-    let result_colored = color_result(&data.result);
-    let run_info = match data.run_id {
-        Some(id) => format!("run #{id}, "),
-        None => String::new(),
-    };
-    let (failed_count, deferred_count) = data.subvolumes.iter().fold((0usize, 0usize), |(f, d), sv| {
-        (f + (!sv.success as usize), d + sv.deferred.len())
-    });
-    let count_suffix = match (failed_count, deferred_count) {
-        (0, 0) => String::new(),
-        (0, d) => format!(" ── ({d} deferred)"),
-        (f, 0) => format!(" ── ({f} failed)"),
-        (f, d) => format!(" ── ({f} failed, {d} deferred)"),
-    };
-    writeln!(
-        out,
-        "{}",
-        format!(
-            "── Urd backup: {result_colored} ── [{run_info}{:.1}s] ──{count_suffix}",
-            data.duration_secs,
-        )
-        .bold()
-    )
-    .ok();
-
-    // ── Executed subvolumes ──────────────────────────────────────────
-    if !data.subvolumes.is_empty() {
-        writeln!(out).ok();
-        for sv in &data.subvolumes {
-            let has_deferred = !sv.deferred.is_empty();
-            let has_sends = !sv.sends.is_empty();
-
-            // Status label: OK (with or without deferred), DEFERRED (only), or FAILED
-            let status = if !sv.success {
-                "FAILED".red().to_string()
-            } else if has_deferred && !has_sends {
-                "DEFERRED".yellow().to_string()
-            } else {
-                "OK".green().to_string()
-            };
-
-            let send_info = format_send_info(&sv.sends);
-            writeln!(
-                out,
-                "  {:<6} {}  [{:.1}s]{}",
-                status,
-                sv.name.bold(),
-                sv.duration_secs,
-                send_info,
-            )
-            .ok();
-
-            for d in &sv.deferred {
-                writeln!(out, "    {} {}", "DEFERRED".yellow(), d.reason).ok();
-                writeln!(out, "    \u{2192} {}", d.suggestion).ok();
-            }
-
-            if !sv.structured_errors.is_empty() {
-                // Render structured errors with layered detail
-                for se in &sv.structured_errors {
-                    writeln!(
-                        out,
-                        "    {} {}: {}",
-                        "ERROR".red(),
-                        se.operation,
-                        se.summary
-                    )
-                    .ok();
-                    writeln!(out, "          Why: {}", se.cause).ok();
-                    if let Some(bytes) = se.bytes_transferred {
-                        writeln!(
-                            out,
-                            "          Transferred {} before failure",
-                            ByteSize(bytes)
-                        )
-                        .ok();
-                    }
-                    if !se.remediation.is_empty() {
-                        writeln!(out, "          What to do:").ok();
-                        for step in &se.remediation {
-                            writeln!(out, "            \u{2022} {step}").ok();
-                        }
-                    }
-                }
-            } else {
-                for err in &sv.errors {
-                    writeln!(out, "    {} {}", "ERROR".red(), err).ok();
-                }
-            }
-        }
-    }
-
-    // ── Skipped sends ────────────────────────────────────────────────
-    render_skipped_block(&data.skipped, &mut out);
-
-    // ── Awareness table ──────────────────────────────────────────────
-    let any_not_protected = data.assessments.iter().any(|a| a.status != "PROTECTED");
-    if any_not_protected {
-        writeln!(out).ok();
-        render_assessment_table(data, &mut out);
-        render_assessment_advisories(data, &mut out);
-    } else if !data.assessments.is_empty() {
-        writeln!(out).ok();
-        writeln!(out, "All subvolumes {}.", "sealed".green()).ok();
-    }
-
-    // ── Warnings ─────────────────────────────────────────────────────
-    if !data.warnings.is_empty() {
-        writeln!(out).ok();
-        for warning in &data.warnings {
-            writeln!(out, "{} {}", "WARNING:".yellow().bold(), warning).ok();
-        }
-    }
-
-    // ── Notes (informational, not warnings) ──────────────────────────
-    // Middle-dot glyph, dimmed, two-space indent. No "NOTE:" label —
-    // the dim rendering signals informational tone without yellow gravity.
-    if !data.notes.is_empty() {
-        writeln!(out).ok();
-        for note in &data.notes {
-            writeln!(out, "  {} {}", "·".dimmed(), note.dimmed()).ok();
-        }
-    }
-
-    // ── Transitions (mythic voice on events) ─────────────────────────
-    render_transitions(&data.transitions, &mut out);
-
-    // ── Next-action suggestion ──────────────────────────────────────
-    let has_failures = data.subvolumes.iter().any(|sv| !sv.success);
-    append_suggestion(&SuggestionContext::Backup { has_failures }, &mut out);
-
-    out
-}
-
-/// Render transition events as brief mythic voice lines.
-/// Each transition gets one line. Empty transitions produce no output.
-fn render_transitions(transitions: &[crate::output::TransitionEvent], out: &mut String) {
-    use crate::output::TransitionEvent;
-
-    if transitions.is_empty() {
-        return;
-    }
-    writeln!(out).ok();
-    for t in transitions {
-        match t {
-            TransitionEvent::ThreadRestored { subvolume, drive } => {
-                writeln!(out, "  {}: thread to {} mended.", subvolume, drive).ok();
-            }
-            TransitionEvent::FirstSendToDrive { subvolume, drive } => {
-                writeln!(out, "  {}: first thread to {} established.", subvolume, drive).ok();
-            }
-            TransitionEvent::AllSealed => {
-                writeln!(out, "  All threads hold.").ok();
-            }
-            TransitionEvent::PromiseRecovered {
-                subvolume,
-                from,
-                to,
-            } => {
-                writeln!(
-                    out,
-                    "  {}: {} \u{2192} {}.",
-                    subvolume,
-                    exposure_label(from),
-                    exposure_label(to),
-                )
-                .ok();
-            }
-        }
-    }
-}
-
-fn format_send_info(sends: &[crate::output::SendSummary]) -> String {
-    if sends.is_empty() {
-        return String::new();
-    }
-    let parts: Vec<String> = sends
-        .iter()
-        .map(|s| {
-            let bytes_info = s
-                .bytes_transferred
-                .map(|b| format!(", {}", ByteSize(b)))
-                .unwrap_or_default();
-            format!("{} \u{2192} {}{}", s.send_type, s.drive, bytes_info)
-        })
-        .collect();
-    format!("  ({})", parts.join("; "))
-}
-
-/// Render skipped subvolumes — absent drives and actionable skips only.
-/// [WAIT] and [OFF] skips are suppressed; the summary line covers the total count.
-fn render_skipped_block(skipped: &[crate::output::SkippedSubvolume], out: &mut String) {
-    if skipped.is_empty() {
-        return;
-    }
-
-    // Collect disconnected drive labels and count their skipped sends.
-    let mut not_mounted_drives: Vec<String> = Vec::new();
-    let mut not_mounted_count = 0usize;
-    // Actionable skips: UUID mismatch, space exceeded, etc. (not WAIT/OFF/drive-not-mounted)
-    let mut actionable_skips: Vec<&crate::output::SkippedSubvolume> = Vec::new();
-
-    for skip in skipped {
-        if let Some(label) = skip
-            .reason
-            .strip_prefix("drive ")
-            .and_then(|r| r.strip_suffix(" not mounted"))
-        {
-            if !not_mounted_drives.contains(&label.to_string()) {
-                not_mounted_drives.push(label.to_string());
-            }
-            not_mounted_count += 1;
-        } else if skip.category != SkipCategory::IntervalNotElapsed
-            && skip.category != SkipCategory::Disabled
-            && skip.category != SkipCategory::LocalOnly
-            && skip.category != SkipCategory::ExternalOnly
-            && skip.category != SkipCategory::Unchanged
-        {
-            actionable_skips.push(skip);
-        }
-    }
-
-    if not_mounted_drives.is_empty() && actionable_skips.is_empty() {
-        return;
-    }
-
-    writeln!(out).ok();
-
-    if !not_mounted_drives.is_empty() {
-        writeln!(
-            out,
-            "  Drives disconnected: {}",
-            not_mounted_drives.join(", "),
-        )
-        .ok();
-        writeln!(out, "    {} send(s) skipped", not_mounted_count).ok();
-    }
-
-    for skip in &actionable_skips {
-        writeln!(
-            out,
-            "  {} {}  {}",
-            skip_tag(&skip.category),
-            skip.name.bold(),
-            skip.reason,
-        )
-        .ok();
-    }
-}
-
-/// Render the awareness assessment table (same layout as status command).
-fn render_assessment_table(data: &BackupSummary, out: &mut String) {
-    // Reuse the same table structure as render_subvolume_table in status rendering.
-    // Build a StatusOutput-compatible view for the shared table formatter.
-    if data.assessments.is_empty() {
-        return;
-    }
-
-    // Collect drive labels from assessments
-    let mut drive_labels: Vec<String> = Vec::new();
-    for assessment in &data.assessments {
-        for ext in &assessment.external {
-            if ext.mounted && !drive_labels.contains(&ext.drive_label) {
-                drive_labels.push(ext.drive_label.clone());
-            }
-        }
-    }
-
-    let has_promises = data.assessments.iter().any(|a| a.promise_level.is_some());
-
-    // Build headers: EXPOSURE  [PROTECTION]  SUBVOLUME  LOCAL  [DRIVE1]  [DRIVE2]
-    let mut headers: Vec<String> = vec!["EXPOSURE".to_string()];
-    if has_promises {
-        // NOTE: Level names (guarded/protected/resilient) stay until Phase 6
-        headers.push("PROTECTION".to_string());
-    }
-    headers.push("SUBVOLUME".to_string());
-    headers.push("LOCAL".to_string());
-    for label in &drive_labels {
-        headers.push(label.clone());
-    }
-
-    // Build rows
-    let mut rows: Vec<Vec<String>> = Vec::new();
-    for assessment in &data.assessments {
-        let mut row = vec![exposure_label(&assessment.status)];
-        if has_promises {
-            row.push(
-                assessment
-                    .promise_level
-                    .clone()
-                    .unwrap_or_else(|| "\u{2014}".to_string()),
-            );
-        }
-        row.push(assessment.name.clone());
-        row.push(assessment.local_snapshot_count.to_string());
-
-        for label in &drive_labels {
-            let count = assessment
-                .external
-                .iter()
-                .find(|e| e.drive_label == *label)
-                .and_then(|e| e.snapshot_count);
-            row.push(match count {
-                Some(c) if c > 0 => c.to_string(),
-                _ => "\u{2014}".to_string(),
-            });
-        }
-
-        rows.push(row);
-    }
-
-    format_status_table(&headers, &rows, Some(0), None, out);
-}
-
-/// Render advisories and errors from awareness assessments.
-fn render_assessment_advisories(data: &BackupSummary, out: &mut String) {
-    for assessment in &data.assessments {
-        for error in &assessment.errors {
-            writeln!(out, "  {} {}: {}", "ERROR".red(), assessment.name, error).ok();
-        }
-        for advisory in &assessment.advisories {
-            writeln!(
-                out,
-                "  {} {}: {}",
-                "NOTE".dimmed(),
-                assessment.name,
-                advisory,
-            )
-            .ok();
-        }
-    }
-}
-
-// ── Get ────────────────────────────────────────────────────────────────
-
-/// Render get metadata according to the given mode (for stderr, not content).
-#[must_use]
-pub fn render_get(data: &GetOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_get_interactive(data),
-        OutputMode::Daemon => render_get_daemon(data),
-    }
-}
-
-fn render_get_daemon(data: &GetOutput) -> String {
-    serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-}
-
-fn render_get_interactive(data: &GetOutput) -> String {
-    let size = ByteSize(data.file_size);
-    format!(
-        "Retrieving from snapshot {} ({}) — {}\n",
-        data.snapshot, data.snapshot_date, size,
-    )
-}
-
-// ── Plan ────────────────────────────────────────────────────────────────
-
-// ── Empty Plan ─────────────────────────────────────────────────────────
-
-/// Render an explanation for why a manual backup produced an empty plan.
-#[must_use]
-pub fn render_empty_plan(explanation: &crate::output::EmptyPlanExplanation) -> String {
-    let mut out = String::new();
-    let reasons = explanation.reasons.join("; ");
-    let _ = write!(out, "Nothing to back up — {reasons}.");
-    if let Some(ref suggestion) = explanation.suggestion {
-        let _ = write!(out, "\n  {suggestion}");
-    }
-    let _ = writeln!(out);
-    out
-}
-
-// ── Pre-Action Summary ─────────────────────────────────────────────────
-
-/// Render a pre-action briefing for manual backup runs.
-#[must_use]
-pub fn render_pre_action(summary: &PreActionSummary) -> String {
-    let mut out = String::new();
-
-    // Build drive list string
-    let drive_labels: Vec<&str> = summary
-        .send_plan
-        .iter()
-        .map(|d| d.drive_label.as_str())
-        .collect();
-    let drive_list = format_list(&drive_labels);
-
-    // Total estimated bytes across all drives
-    let total_bytes: Option<u64> = {
-        let sum: u64 = summary
-            .send_plan
-            .iter()
-            .filter_map(|d| d.estimated_bytes)
-            .sum();
-        if sum > 0 { Some(sum) } else { None }
-    };
-
-    // Size annotation
-    let size_str = total_bytes
-        .map(|b| format!(", ~{}", ByteSize(b)))
-        .unwrap_or_default();
-
-    // Main line depends on filters
-    if summary.filters.local_only {
-        let _ = writeln!(
-            out,
-            "Snapshotting {} subvolume{}.",
-            summary.snapshot_count,
-            if summary.snapshot_count == 1 { "" } else { "s" }
-        );
-    } else if summary.filters.external_only {
-        let total_sends: usize = summary.send_plan.iter().map(|d| d.subvolume_count).sum();
-        let _ = writeln!(
-            out,
-            "Sending to {drive_list}.\n  {total_sends} subvolume{}{size_str}",
-            if total_sends == 1 { "" } else { "s" },
-        );
-    } else if let Some(ref name) = summary.filters.subvolume {
-        let _ = writeln!(
-            out,
-            "Backing up {name} to {drive_list}.\n  1 snapshot{size_str}",
-        );
-    } else {
-        let total_sends: usize = summary.send_plan.iter().map(|d| d.subvolume_count).sum();
-        let _ = writeln!(
-            out,
-            "Backing up everything to {drive_list}.\n  {} snapshot{}, {total_sends} send{}{size_str}",
-            summary.snapshot_count,
-            if summary.snapshot_count == 1 { "" } else { "s" },
-            if total_sends == 1 { "" } else { "s" },
-        );
-    }
-
-    // Disconnected drives
-    for d in &summary.disconnected_drives {
-        match d.role {
-            DriveRole::Offsite => {
-                let _ = writeln!(
-                    out,
-                    "  {} is away — copies will update when it returns.",
-                    d.label
-                );
-            }
-            _ => {
-                let _ = writeln!(out, "  {} not connected.", d.label);
-            }
-        }
-    }
-
-    out
-}
-
-/// Format a list of items as "A", "A and B", or "A, B, and C".
-fn format_list(items: &[&str]) -> String {
-    match items.len() {
-        0 => String::new(),
-        1 => items[0].to_string(),
-        2 => format!("{} and {}", items[0], items[1]),
-        _ => {
-            let (last, rest) = items.split_last().unwrap();
-            format!("{}, and {last}", rest.join(", "))
-        }
-    }
-}
-
-/// Render plan output according to the given mode.
-#[must_use]
-pub fn render_plan(data: &PlanOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_plan_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_plan_interactive(data: &PlanOutput) -> String {
-    let mut out = String::new();
-
-    // Verdict line first (UPI 045 Rule 5 — first line is the answer).
-    // 4-arm match per Finding 1: the zero-subvolume arm prevents a
-    // trust failure where "no subvolumes configured" rendered as
-    // "All sealed." (same well-meaning lie as Finding 1's plan analogue
-    // for `urd doctor` — see R-10).
-    let configured = data.summary.configured_subvolumes;
-    let ops_empty = data.operations.is_empty();
-    let skips_len = data.skipped.len();
-    let op_count = data.operations.len();
-    let verdict_line = match (configured, ops_empty, skips_len) {
-        (0, _, _) => "No subvolumes configured.".dimmed().to_string(),
-        (_, true, 0) => "All sealed.".green().bold().to_string(),
-        (_, true, _) => "No backups planned (all skipped \u{2014} see below).".yellow().to_string(),
-        (_, false, _) => format!("{op_count} operations planned.").bold().to_string(),
-    };
-    writeln!(out, "{verdict_line}").ok();
-    writeln!(out).ok();
-
-    // === Warnings ===
-    if !data.warnings.is_empty() {
-        for warning in &data.warnings {
-            writeln!(out, "  {}  {}", "[WARNING]".yellow().bold(), warning).ok();
-        }
-        writeln!(out).ok();
-    }
-
-    if ops_empty && skips_len == 0 {
-        return out;
-    }
-
-    // === Planned operations ===
-    if !data.operations.is_empty() {
-        writeln!(out, "{}", "=== Planned operations ===".bold()).ok();
-        let mut current_subvol: Option<&str> = None;
-        for entry in &data.operations {
-            if current_subvol != Some(&entry.subvolume) {
-                if current_subvol.is_some() {
-                    writeln!(out).ok();
-                }
-                writeln!(out, "{}:", entry.subvolume.bold()).ok();
-                current_subvol = Some(&entry.subvolume);
-            }
-
-            let label = match entry.operation.as_str() {
-                "create" => "[CREATE]".green().to_string(),
-                "send" => "[SEND]".blue().to_string(),
-                "delete" => "[DELETE]".yellow().to_string(),
-                other => format!("[{other}]"),
-            };
-            let size_annotation = match (entry.estimated_bytes, entry.is_full_send) {
-                (Some(bytes), Some(true)) => format!(" ~{}", ByteSize(bytes)),
-                (Some(bytes), Some(false)) => format!(" last: ~{}", ByteSize(bytes)),
-                _ => String::new(),
-            };
-            writeln!(out, "  {:<10} {}{}", label, entry.detail, size_annotation.dimmed()).ok();
-        }
-        writeln!(out).ok();
-    }
-
-    // === Skipped (N) ===
-    if !data.skipped.is_empty() {
-        writeln!(
-            out,
-            "{}",
-            format!("=== Skipped ({}) ===", data.skipped.len()).dimmed()
-        )
-        .ok();
-        render_plan_skipped_grouped(&data.skipped, &mut out);
-    }
-
-    writeln!(out).ok();
-
-    // Build sends portion of summary, with estimated total if available.
-    let sends_str = if data.summary.sends == 0 {
-        "0 sends".to_string()
-    } else if let Some(total) = data.summary.estimated_total_bytes {
-        let sends_with_estimates = data
-            .operations
-            .iter()
-            .filter(|op| op.operation == "send" && op.estimated_bytes.is_some())
-            .count();
-        if sends_with_estimates == data.summary.sends {
-            format!("{} sends (~{} total)", data.summary.sends, ByteSize(total))
-        } else {
-            format!(
-                "{} sends (~{} estimated for {} of {})",
-                data.summary.sends,
-                ByteSize(total),
-                sends_with_estimates,
-                data.summary.sends
-            )
-        }
-    } else {
-        format!("{} sends", data.summary.sends)
-    };
-
-    writeln!(
-        out,
-        "{}",
-        format!(
-            "Summary: {}, {} snapshots, {} deletions, {} skipped",
-            sends_str,
-            data.summary.snapshots,
-            data.summary.deletions,
-            data.summary.skipped
-        )
-        .bold()
-    )
-    .ok();
-
-    // ── Next-action suggestion ──────────────────────────────────────
-    let has_space_skip = data
-        .skipped
-        .iter()
-        .any(|s| s.category == SkipCategory::SpaceExceeded);
-    append_suggestion(
-        &SuggestionContext::Plan {
-            has_operations: !data.operations.is_empty(),
-            has_space_skip,
-        },
-        &mut out,
-    );
-
-    out
-}
-
-/// Render skipped subvolumes grouped by category for plan output.
-fn render_plan_skipped_grouped(skipped: &[SkippedSubvolume], out: &mut String) {
-    // Collect by category in defined render order.
-    let categories = [
-        SkipCategory::DriveNotMounted,
-        SkipCategory::IntervalNotElapsed,
-        SkipCategory::Disabled,
-        SkipCategory::LocalOnly,
-        SkipCategory::SpaceExceeded,
-        SkipCategory::NoSnapshotsAvailable,
-        SkipCategory::ExternalOnly,
-        SkipCategory::Unchanged,
-        SkipCategory::Other,
-    ];
-
-    for cat in &categories {
-        let items: Vec<&SkippedSubvolume> =
-            skipped.iter().filter(|s| &s.category == cat).collect();
-        if items.is_empty() {
-            continue;
-        }
-        match cat {
-            SkipCategory::DriveNotMounted => render_drive_not_mounted_group(&items, out),
-            SkipCategory::IntervalNotElapsed => render_interval_group(&items, out),
-            SkipCategory::Disabled => render_named_group(&items, cat, "Disabled", out),
-            SkipCategory::LocalOnly => render_named_group(&items, cat, "Local only", out),
-            SkipCategory::ExternalOnly => render_named_group(&items, cat, "External only", out),
-            SkipCategory::Unchanged
-            | SkipCategory::SpaceExceeded
-            | SkipCategory::NoSnapshotsAvailable
-            | SkipCategory::Other => {
-                render_individual_skips(&items, cat, out);
-            }
-        }
-    }
-}
-
-/// Render DriveNotMounted skips, sub-grouped by drive label with subvolume counts.
-fn render_drive_not_mounted_group(items: &[&SkippedSubvolume], out: &mut String) {
-    // Extract drive label from reason: "drive {label} not mounted"
-    let mut drives: Vec<(String, usize)> = Vec::new();
-    for item in items {
-        let label = item
-            .reason
-            .strip_prefix("drive ")
-            .and_then(|r| r.strip_suffix(" not mounted"))
-            .unwrap_or("unknown")
-            .to_string();
-        if let Some(entry) = drives.iter_mut().find(|(l, _)| *l == label) {
-            entry.1 += 1;
-        } else {
-            drives.push((label, 1));
-        }
-    }
-    let parts: Vec<String> = drives
-        .iter()
-        .map(|(label, count)| {
-            let noun = if *count == 1 { "subvolume" } else { "subvolumes" };
-            format!("{label} ({count} {noun})")
-        })
-        .collect();
-    writeln!(
-        out,
-        "  {}  {} {}",
-        skip_tag(&SkipCategory::DriveNotMounted),
-        "Disconnected:".dimmed(),
-        parts.join(", "),
-    )
-    .ok();
-}
-
-/// Render IntervalNotElapsed skips as a single line with count and shortest duration.
-fn render_interval_group(items: &[&SkippedSubvolume], out: &mut String) {
-    let shortest = items
-        .iter()
-        .filter_map(|s| parse_duration_to_minutes(&s.reason))
-        .min();
-
-    let suffix = if let Some(mins) = shortest {
-        format!(" (next in ~{})", format_duration_short(mins as i64))
-    } else {
-        String::new()
-    };
-
-    writeln!(
-        out,
-        "  {}  {} {} subvolumes{}",
-        skip_tag(&SkipCategory::IntervalNotElapsed),
-        "Interval not elapsed:".dimmed(),
-        items.len(),
-        suffix,
-    )
-    .ok();
-}
-
-/// Render a skip group as: `[TAG]  Label: name1, name2`.
-fn render_named_group(
-    items: &[&SkippedSubvolume],
-    category: &SkipCategory,
-    label: &str,
-    out: &mut String,
-) {
-    let names: Vec<&str> = items.iter().map(|s| s.name.as_str()).collect();
-    writeln!(
-        out,
-        "  {}  {} {}",
-        skip_tag(category),
-        format!("{label}:").dimmed(),
-        names.join(", "),
-    )
-    .ok();
-}
-
-/// Map skip category to a colored tag for display.
-fn skip_tag(category: &SkipCategory) -> String {
+/// Map a skip category to its colored display tag.
+///
+/// `pub(super)` for sibling voice/* sub-modules (plan.rs, backup.rs) — the
+/// `[TAG]` chips show up identically wherever skips are listed, so the
+/// canonical lookup belongs at the parent.
+pub(super) fn skip_tag(category: &SkipCategory) -> String {
     match category {
         SkipCategory::SpaceExceeded => "[SPACE]".yellow().to_string(),
         SkipCategory::IntervalNotElapsed => "[WAIT]".dimmed().to_string(),
@@ -967,186 +252,11 @@ fn skip_tag(category: &SkipCategory) -> String {
     }
 }
 
-/// Render SpaceExceeded or Other skips as individual lines (detail matters).
-fn render_individual_skips(
-    items: &[&SkippedSubvolume],
-    category: &SkipCategory,
-    out: &mut String,
-) {
-    let tag = skip_tag(category);
-    for item in items {
-        writeln!(out, "  {} {}: {}", tag, item.name, item.reason.dimmed()).ok();
-    }
-}
-
-// ── Events ──────────────────────────────────────────────────────────────
-
-/// Render the `urd events` view. Delegates to `voice_events` for the
-/// per-variant columnar / NDJSON formatting.
-#[must_use]
-pub fn render_events(view: &crate::output::EventsView, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => crate::voice_events::render_interactive(view),
-        OutputMode::Daemon => crate::voice_events::render_ndjson(view),
-    }
-}
-
-// ── History ─────────────────────────────────────────────────────────────
-
-/// Render history (recent runs) output.
-#[must_use]
-pub fn render_history(data: &HistoryOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_history_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_history_interactive(data: &HistoryOutput) -> String {
-    let mut out = String::new();
-
-    if data.runs.is_empty() {
-        writeln!(out, "{}", "No backup runs recorded.".dimmed()).ok();
-        return out;
-    }
-
-    let headers = vec![
-        "RUN".to_string(),
-        "STARTED".to_string(),
-        "MODE".to_string(),
-        "RESULT".to_string(),
-        "DURATION".to_string(),
-    ];
-    let rows: Vec<Vec<String>> = data
-        .runs
-        .iter()
-        .map(|r| {
-            vec![
-                r.id.to_string(),
-                r.started_at.clone(),
-                r.mode.clone(),
-                r.result.clone(),
-                r.duration.clone().unwrap_or_else(|| "running".to_string()),
-            ]
-        })
-        .collect();
-    format_history_table(&headers, &rows, &mut out);
-
-    out
-}
-
-/// Render subvolume history output.
-#[must_use]
-pub fn render_subvolume_history(data: &SubvolumeHistoryOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_subvolume_history_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_subvolume_history_interactive(data: &SubvolumeHistoryOutput) -> String {
-    let mut out = String::new();
-
-    if data.operations.is_empty() {
-        writeln!(
-            out,
-            "No operations recorded for subvolume {:?}.",
-            data.subvolume
-        )
-        .ok();
-        return out;
-    }
-
-    writeln!(out, "{}", format!("History for {}:", data.subvolume).bold()).ok();
-    writeln!(out).ok();
-
-    let headers = vec![
-        "RUN".to_string(),
-        "OPERATION".to_string(),
-        "DRIVE".to_string(),
-        "RESULT".to_string(),
-        "DURATION".to_string(),
-        "ERROR".to_string(),
-    ];
-    let rows: Vec<Vec<String>> = data
-        .operations
-        .iter()
-        .map(|op| {
-            vec![
-                op.run_id.to_string(),
-                op.operation.clone(),
-                op.drive.clone().unwrap_or_else(|| "\u{2014}".to_string()),
-                op.result.clone(),
-                op.duration
-                    .clone()
-                    .unwrap_or_else(|| "\u{2014}".to_string()),
-                truncate_str(op.error.as_deref().unwrap_or(""), 30),
-            ]
-        })
-        .collect();
-    format_history_table(&headers, &rows, &mut out);
-
-    out
-}
-
-/// Render failures output.
-#[must_use]
-pub fn render_failures(data: &FailuresOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_failures_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_failures_interactive(data: &FailuresOutput) -> String {
-    let mut out = String::new();
-
-    if data.failures.is_empty() {
-        writeln!(out, "{}", "No failures recorded.".green()).ok();
-        return out;
-    }
-
-    writeln!(
-        out,
-        "{}",
-        format!("{} failure(s):", data.failures.len()).red().bold()
-    )
-    .ok();
-    writeln!(out).ok();
-
-    let headers = vec![
-        "RUN".to_string(),
-        "SUBVOLUME".to_string(),
-        "OPERATION".to_string(),
-        "DRIVE".to_string(),
-        "ERROR".to_string(),
-    ];
-    let rows: Vec<Vec<String>> = data
-        .failures
-        .iter()
-        .map(|f| {
-            vec![
-                f.run_id.to_string(),
-                f.subvolume.clone(),
-                f.operation.clone(),
-                f.drive.clone().unwrap_or_else(|| "\u{2014}".to_string()),
-                truncate_str(f.error.as_deref().unwrap_or("unknown"), 40),
-            ]
-        })
-        .collect();
-    format_history_table(&headers, &rows, &mut out);
-
-    out
-}
-
 /// Format a table with result-colored RESULT column.
-fn format_history_table(headers: &[String], rows: &[Vec<String>], out: &mut String) {
+///
+/// `pub(super)` for sibling voice/* sub-modules (history.rs, verify.rs) per
+/// UPI 050 phase 2 — cross-renderer helper, single definition stays here.
+pub(super) fn format_history_table(headers: &[String], rows: &[Vec<String>], out: &mut String) {
     let cols = headers.len();
     let mut widths: Vec<usize> = headers.iter().map(|h| h.len()).collect();
     for row in rows {
@@ -1187,7 +297,11 @@ fn format_history_table(headers: &[String], rows: &[Vec<String>], out: &mut Stri
     }
 }
 
-fn truncate_str(s: &str, max_len: usize) -> String {
+/// Truncate a string to a maximum visible length, appending an ellipsis when
+/// trimmed. Char-boundary-safe.
+///
+/// `pub(super)` for sibling voice/* sub-modules (history.rs, verify.rs).
+pub(super) fn truncate_str(s: &str, max_len: usize) -> String {
     if s.len() <= max_len {
         return s.to_string();
     }
@@ -1200,893 +314,7 @@ fn truncate_str(s: &str, max_len: usize) -> String {
     format!("{}...", &s[..end])
 }
 
-// ── Calibrate ───────────────────────────────────────────────────────────
 
-/// Render calibrate output according to the given mode.
-#[must_use]
-pub fn render_calibrate(data: &CalibrateOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_calibrate_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_calibrate_interactive(data: &CalibrateOutput) -> String {
-    let mut out = String::new();
-
-    writeln!(
-        out,
-        "{}",
-        "Urd calibrate \u{2014} measuring snapshot sizes".bold()
-    )
-    .ok();
-    writeln!(out).ok();
-
-    for entry in &data.entries {
-        match &entry.result {
-            CalibrateResult::Ok { snapshot, bytes } => {
-                writeln!(
-                    out,
-                    "  {} ({}) {}",
-                    entry.name.bold(),
-                    snapshot,
-                    ByteSize(*bytes),
-                )
-                .ok();
-            }
-            CalibrateResult::Skipped { reason } => {
-                writeln!(out, "  {} {} ({})", "SKIP".dimmed(), entry.name, reason).ok();
-            }
-            CalibrateResult::Failed { snapshot, error } => {
-                writeln!(
-                    out,
-                    "  {} ({}) {}",
-                    entry.name.bold(),
-                    snapshot,
-                    "FAILED".red(),
-                )
-                .ok();
-                writeln!(out, "    {error}").ok();
-            }
-        }
-    }
-
-    writeln!(out).ok();
-    writeln!(
-        out,
-        "Calibrated {} subvolume(s), skipped {}.",
-        data.calibrated, data.skipped
-    )
-    .ok();
-    writeln!(
-        out,
-        "Sizes stored in state database. The planner will use these as fallback"
-    )
-    .ok();
-    writeln!(out, "estimates when no send history exists.").ok();
-
-    out
-}
-
-// ── Verify ──────────────────────────────────────────────────────────────
-
-/// Render verify output according to the given mode.
-#[must_use]
-pub fn render_verify(data: &VerifyOutput, mode: OutputMode, detail: bool) -> String {
-    match mode {
-        OutputMode::Interactive => render_verify_interactive(data, detail),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_verify_interactive(data: &VerifyOutput, detail: bool) -> String {
-    if detail {
-        return render_verify_detail(data);
-    }
-    render_verify_findings_first(data)
-}
-
-/// Detail mode: every check for every subvolume/drive (original verbose output).
-fn render_verify_detail(data: &VerifyOutput) -> String {
-    let mut out = String::new();
-
-    for sv in &data.subvolumes {
-        writeln!(out, "Verifying {}...", sv.name.bold()).ok();
-
-        for drive in &sv.drives {
-            writeln!(out, "  {}:", drive.label.bold()).ok();
-
-            for check in &drive.checks {
-                let status_str = match check.status.as_str() {
-                    "ok" => format!("{}   ", "OK".green()),
-                    "warn" => format!("{}  ", "WARN".yellow()),
-                    "fail" => format!("{}  ", "FAIL".red()),
-                    other => format!("{other:<6}"),
-                };
-                let detail = check.detail.as_deref().unwrap_or(&check.name);
-                writeln!(out, "    {status_str} {detail}").ok();
-            }
-        }
-
-        writeln!(out).ok();
-    }
-
-    render_verify_tail(data, &mut out);
-    out
-}
-
-/// Findings-first mode: problems first, noise collapsed.
-fn render_verify_findings_first(data: &VerifyOutput) -> String {
-    let mut out = String::new();
-
-    let (findings, absent_drives) = classify_verify_checks(data);
-
-    // Render findings grouped by subvolume/drive
-    if findings.is_empty() {
-        writeln!(
-            out,
-            "{}",
-            format!(
-                "All threads intact. {} verified, {} OK.",
-                pluralize(data.subvolumes.len(), "subvolume", "subvolumes"),
-                pluralize(data.ok_count as usize, "check", "checks")
-            )
-            .green()
-            .bold()
-        )
-        .ok();
-    } else {
-        for (sv_name, drive_label, check) in &findings {
-            let status_str = match check.status.as_str() {
-                "warn" => format!("{}  ", "WARN".yellow()),
-                "fail" => format!("{}  ", "FAIL".red()),
-                other => format!("{other:<6}"),
-            };
-            let detail = check.detail.as_deref().unwrap_or(&check.name);
-            writeln!(out, "{sv_name}/{drive_label}:").ok();
-            writeln!(out, "  {status_str}{detail}").ok();
-            if let Some(ref suggestion) = check.suggestion {
-                writeln!(out, "  \u{2192} {suggestion}").ok();
-            }
-            writeln!(out).ok();
-        }
-
-        // ok_count from verify.rs includes all OK checks; drive-mounted
-        // warnings are counted separately in the absent-drives line below.
-        writeln!(
-            out,
-            "{}",
-            format!(
-                "{} verified, {} OK.",
-                pluralize(data.subvolumes.len(), "subvolume", "subvolumes"),
-                pluralize(data.ok_count as usize, "check", "checks")
-            )
-            .dimmed()
-        )
-        .ok();
-    }
-
-    // Absent drives summary
-    if !absent_drives.is_empty() {
-        writeln!(
-            out,
-            "{}",
-            format!(
-                "{} not mounted ({}) \u{2014} skipped.",
-                pluralize(absent_drives.len(), "drive", "drives"),
-                absent_drives.join(", ")
-            )
-            .dimmed()
-        )
-        .ok();
-    }
-
-    render_verify_tail(data, &mut out);
-    out
-}
-
-/// Shared tail: preflight warnings + next-action suggestion.
-fn render_verify_tail(data: &VerifyOutput, out: &mut String) {
-    // Preflight warnings
-    if !data.preflight_warnings.is_empty() {
-        writeln!(out).ok();
-        writeln!(out, "{}", "Config consistency:".bold()).ok();
-        for warning in &data.preflight_warnings {
-            writeln!(out, "  {} {}", "WARN".yellow(), warning).ok();
-        }
-    }
-
-    // Next-action suggestion
-    append_suggestion(
-        &SuggestionContext::Verify { has_broken: data.fail_count > 0 },
-        out,
-    );
-}
-
-// ── Init ────────────────────────────────────────────────────────────────
-
-/// Render init output according to the given mode.
-#[must_use]
-pub fn render_init(data: &InitOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_init_interactive(data),
-        OutputMode::Daemon => render_init_daemon(data),
-    }
-}
-
-fn render_init_daemon(data: &InitOutput) -> String {
-    serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-}
-
-fn render_init_interactive(data: &InitOutput) -> String {
-    let mut out = String::new();
-
-    writeln!(out, "{}", "Urd initialization".bold()).ok();
-    writeln!(out).ok();
-
-    // ── Infrastructure checks ──────────────────────────────────────
-    for check in &data.infrastructure {
-        let status = format_init_status(check.status);
-        if let Some(ref detail) = check.detail {
-            writeln!(out, "{status} {}: {detail}", check.name).ok();
-        } else {
-            writeln!(out, "{status} {}", check.name).ok();
-        }
-    }
-
-    // ── Subvolume sources ──────────────────────────────────────────
-    writeln!(out).ok();
-    writeln!(out, "{}", "Checking subvolume sources:".bold()).ok();
-    for check in &data.subvolume_sources {
-        let status = format_init_status(check.status);
-        if let Some(ref detail) = check.detail {
-            writeln!(out, "  {status} {}: {detail}", check.name).ok();
-        } else {
-            writeln!(out, "  {status} {}", check.name).ok();
-        }
-    }
-
-    // ── Snapshot roots ─────────────────────────────────────────────
-    writeln!(out).ok();
-    writeln!(out, "{}", "Checking snapshot roots:".bold()).ok();
-    for check in &data.snapshot_roots {
-        let status = format_init_status(check.status);
-        writeln!(out, "  {status} {}", check.name).ok();
-    }
-
-    // ── Drives ─────────────────────────────────────────────────────
-    writeln!(out).ok();
-    writeln!(out, "{}", "Drive status:".bold()).ok();
-    for drive in &data.drives {
-        let status = if drive.mounted {
-            "CONNECTED".green().to_string()
-        } else if drive.role == DriveRole::Offsite {
-            "AWAY".yellow().to_string()
-        } else {
-            "DISCONNECTED".yellow().to_string()
-        };
-        let free_info = drive
-            .free_bytes
-            .map(|b| format!(" ({} free)", ByteSize(b)))
-            .unwrap_or_default();
-        writeln!(
-            out,
-            "  {status} {} [{}] at {}{free_info}",
-            drive.label.bold(),
-            drive.role,
-            drive.mount_path,
-        )
-        .ok();
-    }
-
-    // ── Pin files ──────────────────────────────────────────────────
-    writeln!(out).ok();
-    writeln!(out, "{}", "Pin file status:".bold()).ok();
-    for pin in &data.pin_files {
-        match pin.status {
-            InitStatus::Ok => {
-                writeln!(
-                    out,
-                    "  {} {}/{}: {}",
-                    "OK".green(),
-                    pin.subvolume,
-                    pin.drive,
-                    pin.snapshot_name.as_deref().unwrap_or("—"),
-                )
-                .ok();
-            }
-            InitStatus::Warn => {
-                writeln!(
-                    out,
-                    "  {} {}/{}: no pin file",
-                    "—".dimmed(),
-                    pin.subvolume,
-                    pin.drive,
-                )
-                .ok();
-            }
-            InitStatus::Error => {
-                writeln!(
-                    out,
-                    "  {} {}/{}: {}",
-                    "ERROR".red(),
-                    pin.subvolume,
-                    pin.drive,
-                    pin.error.as_deref().unwrap_or("unknown error"),
-                )
-                .ok();
-            }
-        }
-    }
-
-    // ── Incomplete snapshots ───────────────────────────────────────
-    if !data.incomplete_snapshots.is_empty() {
-        writeln!(out).ok();
-        writeln!(
-            out,
-            "{}",
-            "Potentially incomplete snapshots on external drives:".bold()
-        )
-        .ok();
-        for inc in &data.incomplete_snapshots {
-            writeln!(
-                out,
-                "  {} {} on {} (not pinned, may be from interrupted transfer)",
-                "WARNING".yellow(),
-                inc.snapshot,
-                inc.drive,
-            )
-            .ok();
-        }
-    }
-
-    // ── Snapshot counts ────────────────────────────────────────────
-    writeln!(out).ok();
-    writeln!(out, "{}", "Snapshot counts:".bold()).ok();
-    for sc in &data.snapshot_counts {
-        let ext_display = if sc.external_counts.is_empty() {
-            "no drives mounted".to_string()
-        } else {
-            sc.external_counts
-                .iter()
-                .map(|(label, count)| format!("{label}:{count}"))
-                .collect::<Vec<_>>()
-                .join(", ")
-        };
-        writeln!(
-            out,
-            "  {} — local: {}, external: [{ext_display}]",
-            sc.subvolume.bold(),
-            sc.local_count,
-        )
-        .ok();
-    }
-
-    // ── Preflight warnings ─────────────────────────────────────────
-    if !data.preflight_warnings.is_empty() {
-        writeln!(out).ok();
-        writeln!(out, "{}", "Config consistency checks:".bold()).ok();
-        for warning in &data.preflight_warnings {
-            writeln!(out, "  {} {warning}", "WARN".yellow()).ok();
-        }
-    }
-
-    writeln!(out).ok();
-    writeln!(out, "{}", "Initialization complete.".green().bold()).ok();
-
-    out
-}
-
-fn format_init_status(status: InitStatus) -> String {
-    match status {
-        InitStatus::Ok => "OK".green().to_string(),
-        InitStatus::Warn => "WARN".yellow().to_string(),
-        InitStatus::Error => "ERROR".red().to_string(),
-    }
-}
-
-// ── Sentinel Status ──────────────────────────────────────────────────────
-
-/// Render sentinel status output according to the given mode.
-#[must_use]
-pub fn render_sentinel_status(data: &SentinelStatusOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_sentinel_status_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data)
-                .unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_sentinel_status_interactive(data: &SentinelStatusOutput) -> String {
-    let mut out = String::new();
-
-    match data {
-        SentinelStatusOutput::Running { state, uptime } => {
-            writeln!(out, "{}", "SENTINEL — watching".bold()).ok();
-            writeln!(out).ok();
-            writeln!(
-                out,
-                "  {:<14}since {} (PID {})",
-                "Running", uptime, state.pid
-            )
-            .ok();
-
-            // Assessment timing
-            if let Some(ref last) = state.last_assessment {
-                let tick_desc = format_tick_description(state.tick_interval_secs, &state.promise_states);
-                writeln!(out, "  {:<14}{} (tick: {})", "Assessment", last, tick_desc).ok();
-            }
-
-            // Mounted drives
-            if state.mounted_drives.is_empty() {
-                writeln!(out, "  {:<14}{}", "Connected", "none".dimmed()).ok();
-            } else {
-                writeln!(out, "  {:<14}{}", "Connected", state.mounted_drives.join(", ")).ok();
-            }
-        }
-        SentinelStatusOutput::NotRunning { last_seen } => {
-            if let Some(seen) = last_seen {
-                writeln!(
-                    out,
-                    "{}",
-                    format!("SENTINEL — not running (last seen {seen})").bold()
-                )
-                .ok();
-            } else {
-                writeln!(out, "{}", "SENTINEL — not running".bold()).ok();
-            }
-            writeln!(out).ok();
-            writeln!(out, "  Start with: {}", "systemctl --user start urd-sentinel".dimmed()).ok();
-            writeln!(out, "  Or: {}", "urd sentinel run".dimmed()).ok();
-        }
-    }
-
-    out
-}
-
-fn format_tick_description(tick_secs: u64, promise_states: &[crate::output::SentinelPromiseState]) -> String {
-    let tick_str = if tick_secs >= 60 {
-        format!("{}m", tick_secs / 60)
-    } else {
-        format!("{tick_secs}s")
-    };
-
-    let worst = promise_states
-        .iter()
-        .map(|p| p.status.as_str())
-        .min_by_key(|s| match *s {
-            "UNPROTECTED" => 0,
-            "AT RISK" => 1,
-            "PROTECTED" => 2,
-            _ => 0,
-        });
-
-    let state_desc = match worst {
-        Some("PROTECTED") | None => "all promises held",
-        Some("AT RISK") => "promises at risk",
-        Some("UNPROTECTED") => "promises broken",
-        Some(_) => "assessing",
-    };
-
-    format!("{tick_str} — {state_desc}")
-}
-
-// ── Emergency ─────────────────────────────────────────────────────────
-
-/// Render the emergency assessment (before user confirms).
-#[must_use]
-pub fn render_emergency(data: &EmergencyOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_emergency_interactive(data),
-        OutputMode::Daemon => serde_json::to_string_pretty(data).unwrap_or_default(),
-    }
-}
-
-fn render_emergency_interactive(data: &EmergencyOutput) -> String {
-    let mut out = String::new();
-
-    if !data.has_crisis() {
-        writeln!(out).ok();
-        writeln!(
-            out,
-            "{}",
-            "No crisis detected.".green()
-        )
-        .ok();
-        writeln!(out).ok();
-        for root in &data.roots {
-            let free = ByteSize(root.free_bytes);
-            let status = match root.min_free_bytes {
-                Some(threshold) => format!(
-                    "{} free (threshold: {})  {}",
-                    free,
-                    ByteSize(threshold),
-                    "OK".green()
-                ),
-                None => format!("{} free (no threshold configured)", free),
-            };
-            writeln!(out, "  {}  — {}", root.root.display(), status).ok();
-        }
-        return out;
-    }
-
-    writeln!(out).ok();
-    writeln!(out, "{}", "Urd sees a crisis.".red().bold()).ok();
-
-    for root in &data.roots {
-        if !root.is_critical {
-            continue;
-        }
-
-        let free = ByteSize(root.free_bytes);
-        let threshold = ByteSize(root.min_free_bytes.unwrap_or(0));
-        let total_snaps: usize = root.subvolumes.iter().map(|s| s.snapshot_count).sum();
-        let total_delete: usize = root.subvolumes.iter().map(|s| s.delete_count).sum();
-        let total_keep: usize = root.subvolumes.iter().map(|s| s.keep_count).sum();
-        let total_pinned: usize = root.subvolumes.iter().map(|s| s.pinned_count).sum();
-
-        writeln!(out).ok();
-        writeln!(
-            out,
-            "{} — {} free (threshold: {})",
-            root.root.display().to_string().bold(),
-            free.to_string().red(),
-            threshold
-        )
-        .ok();
-        writeln!(
-            out,
-            "  {} snapshots across {} subvolumes",
-            total_snaps,
-            root.subvolumes.len()
-        )
-        .ok();
-
-        // Per-subvolume detail
-        for sv in &root.subvolumes {
-            writeln!(
-                out,
-                "    {}: {} snapshots, keep {}, delete {}",
-                sv.name, sv.snapshot_count, sv.keep_count, sv.delete_count
-            )
-            .ok();
-        }
-
-        writeln!(out, "  Chain parents pinned: {}", total_pinned).ok();
-
-        // Unsent snapshot advisory (F3)
-        if root.unsent_count > 0 {
-            writeln!(out).ok();
-            writeln!(
-                out,
-                "  {} unsent snapshots will be deleted.",
-                root.unsent_count
-            )
-            .ok();
-            if !root.drives_needing_full_send.is_empty() {
-                writeln!(
-                    out,
-                    "  Next send to {} will be a full send.",
-                    root.drives_needing_full_send.join(", ")
-                )
-                .ok();
-            }
-        }
-
-        writeln!(out).ok();
-        writeln!(
-            out,
-            "  This will delete {} snapshots.",
-            total_delete.to_string().yellow()
-        )
-        .ok();
-        writeln!(
-            out,
-            "  Your newest snapshot and all chain parents will be preserved."
-        )
-        .ok();
-        writeln!(out, "  {} snapshots will remain.", total_keep).ok();
-    }
-
-    // Show non-critical roots
-    let non_critical: Vec<_> = data.roots.iter().filter(|r| !r.is_critical).collect();
-    if !non_critical.is_empty() {
-        writeln!(out).ok();
-        for root in non_critical {
-            let free = ByteSize(root.free_bytes);
-            let status = match root.min_free_bytes {
-                Some(threshold) => format!(
-                    "{} free (threshold: {})  {}",
-                    free,
-                    ByteSize(threshold),
-                    "OK".green()
-                ),
-                None => format!("{} free (no threshold configured)", free),
-            };
-            writeln!(out, "  {}  — {}", root.root.display(), status).ok();
-        }
-    }
-
-    writeln!(out).ok();
-    out
-}
-
-/// Render the result of emergency deletions.
-#[must_use]
-pub fn render_emergency_result(data: &EmergencyResult, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_emergency_result_interactive(data),
-        OutputMode::Daemon => serde_json::to_string_pretty(data).unwrap_or_default(),
-    }
-}
-
-fn render_emergency_result_interactive(data: &EmergencyResult) -> String {
-    let mut out = String::new();
-
-    writeln!(out).ok();
-    if data.failed == 0 {
-        writeln!(
-            out,
-            "Freed {}. {} snapshots remain in {}.",
-            ByteSize(data.freed_bytes).to_string().green(),
-            data.remaining_snapshots,
-            data.root.display()
-        )
-        .ok();
-    } else {
-        writeln!(
-            out,
-            "Deleted {} snapshots ({} failed). Freed {}.",
-            data.deleted,
-            data.failed.to_string().red(),
-            ByteSize(data.freed_bytes)
-        )
-        .ok();
-        writeln!(out, "{} snapshots remain.", data.remaining_snapshots).ok();
-    }
-
-    writeln!(
-        out,
-        "{} now has {} free.",
-        data.root.display(),
-        ByteSize(data.remaining_free)
-    )
-    .ok();
-
-    if data.still_critical {
-        writeln!(out).ok();
-        writeln!(
-            out,
-            "{}",
-            "Still below threshold. Only pinned and latest snapshots remain."
-                .yellow()
-        )
-        .ok();
-        writeln!(
-            out,
-            "Manual intervention may be needed."
-        )
-        .ok();
-    }
-
-    out
-}
-
-
-// ── Retention Preview ─────────────────────────────────────────────────
-
-/// Render retention preview output.
-#[must_use]
-pub fn render_retention_preview(data: &RetentionPreviewOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_retention_preview_interactive(data),
-        OutputMode::Daemon => serde_json::to_string_pretty(data)
-            .unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}")),
-    }
-}
-
-fn render_retention_preview_interactive(data: &RetentionPreviewOutput) -> String {
-    let mut out = String::new();
-
-    for (i, preview) in data.previews.iter().enumerate() {
-        if i > 0 {
-            writeln!(out).ok();
-        }
-
-        writeln!(
-            out,
-            "{}",
-            format!("Retention preview for \"{}\":", preview.subvolume_name).bold()
-        )
-        .ok();
-        writeln!(out, "  Policy: {}", preview.policy_description).ok();
-        writeln!(out, "  Snapshot interval: {}", preview.snapshot_interval).ok();
-
-        if preview.recovery_windows.is_empty() {
-            writeln!(out).ok();
-            writeln!(out, "  Recovery windows: {}", "none".yellow()).ok();
-            writeln!(
-                out,
-                "    No local recovery. External drive must be connected to restore."
-            )
-            .ok();
-            writeln!(
-                out,
-                "    Only the current incremental chain parent is kept locally (1 snapshot)."
-            )
-            .ok();
-        } else {
-            writeln!(out).ok();
-            writeln!(out, "  Recovery windows (cumulative):").ok();
-            for w in &preview.recovery_windows {
-                writeln!(
-                    out,
-                    "    {:8} {}",
-                    format!("{}:", w.granularity).dimmed(),
-                    w.cumulative_description
-                )
-                .ok();
-            }
-        }
-
-        if let Some(ref estimate) = preview.estimated_disk_usage {
-            writeln!(out).ok();
-            writeln!(
-                out,
-                "  Estimated snapshots: {} ({})",
-                estimate.total_count,
-                format_snapshot_breakdown(&preview.recovery_windows)
-            )
-            .ok();
-            writeln!(
-                out,
-                "  Estimated disk usage: ~{} ({} snapshots x ~{} average)",
-                ByteSize(estimate.total_bytes),
-                estimate.total_count,
-                ByteSize(estimate.per_snapshot_bytes)
-            )
-            .ok();
-            writeln!(
-                out,
-                "    {}",
-                "Upper bound only. BTRFS shares unchanged data between snapshots;"
-                    .dimmed()
-            )
-            .ok();
-            writeln!(
-                out,
-                "    {}",
-                "actual usage depends on your rate of change and is often 5-10x lower."
-                    .dimmed()
-            )
-            .ok();
-        }
-
-        if let Some(ref comparison) = preview.transient_comparison {
-            writeln!(out).ok();
-            let count_diff =
-                comparison.graduated_count.saturating_sub(comparison.transient_count);
-            if let Some(savings) = comparison.savings_bytes {
-                writeln!(
-                    out,
-                    "  Compared to transient: saves ~{} ({} fewer snapshots)",
-                    ByteSize(savings),
-                    count_diff
-                )
-                .ok();
-            } else {
-                writeln!(
-                    out,
-                    "  Compared to transient: saves {} snapshots",
-                    count_diff
-                )
-                .ok();
-            }
-            writeln!(out, "  Loses: {}", comparison.lost_window).ok();
-        }
-    }
-
-    out
-}
-
-fn format_snapshot_breakdown(windows: &[RecoveryWindow]) -> String {
-    windows
-        .iter()
-        .map(|w| format!("{} {}", w.count, w.granularity))
-        .collect::<Vec<_>>()
-        .join(" + ")
-}
-
-
-// ── Default status (bare `urd`) ────────────────────────────────────────
-
-/// Render bare `urd` one-sentence status.
-#[must_use]
-pub fn render_default_status(data: &DefaultStatusOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_default_status_interactive(data),
-        OutputMode::Daemon => render_default_status_daemon(data),
-    }
-}
-
-fn render_default_status_daemon(data: &DefaultStatusOutput) -> String {
-    serde_json::to_string_pretty(data).unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-}
-
-fn render_default_status_interactive(data: &DefaultStatusOutput) -> String {
-    let mut out = String::new();
-
-    // Safety line
-    if data.sealed_count() == data.total {
-        write!(out, "{}", "All connected drives are sealed.".green()).ok();
-    } else {
-        write!(out, "{} of {} sealed.", data.sealed_count(), data.total).ok();
-        if !data.exposed_names.is_empty() {
-            write!(out, " {} {}.", data.exposed_names.join(", "), "exposed".red()).ok();
-        }
-        if !data.waning_names.is_empty() {
-            write!(out, " {} waning.", data.waning_names.join(", ")).ok();
-        }
-    }
-
-    // Health degradation
-    let health_issues = data.degraded_count + data.blocked_count;
-    if health_issues > 0 {
-        let mut parts = Vec::new();
-        if data.blocked_count > 0 {
-            parts.push(format!("{} blocked", data.blocked_count));
-        }
-        if data.degraded_count > 0 {
-            parts.push(format!("{} degraded", data.degraded_count));
-        }
-        write!(out, " {}.", parts.join(", ")).ok();
-    }
-
-    // Last backup age (pre-computed by command handler to keep voice pure)
-    if let Some(age_secs) = data.last_run_age_secs {
-        write!(out, " Last backup {} ago.", humanize_duration(age_secs)).ok();
-    }
-
-    writeln!(out).ok();
-
-    // Next-action suggestion
-    if let Some(ref advice) = data.best_advice {
-        if data.total_needing_attention == 1 {
-            if let Some(ref cmd) = advice.command {
-                writeln!(out, "{}", format!("Run `{cmd}`.").dimmed()).ok();
-            } else if let Some(ref reason) = advice.reason {
-                writeln!(out, "{}", reason.dimmed()).ok();
-            }
-        } else {
-            writeln!(out, "{}", format!("{} subvolumes need attention — run `urd status` for details.", data.total_needing_attention).dimmed()).ok();
-        }
-    } else if data.sealed_count() < data.total || health_issues > 0 {
-        append_suggestion(&SuggestionContext::Default { has_issues: true }, &mut out);
-    } else {
-        writeln!(out, "{}", "Run `urd status` for details, `urd --help` for commands.".dimmed())
-            .ok();
-    }
-
-    out
-}
-
-/// Render first-time message (no config found).
-#[must_use]
-pub fn render_first_time(mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => {
-            "Urd is not configured yet.\nRun `urd init` to get started, or see `urd --help`.\n"
-                .to_string()
-        }
-        OutputMode::Daemon => r#"{"status":"not_configured"}"#.to_string(),
-    }
-}
 
 // ── Staleness Escalation (4a) ──────────────────────────────────────────
 
@@ -2249,196 +477,6 @@ fn append_suggestion(context: &SuggestionContext, out: &mut String) {
         writeln!(out).ok();
         writeln!(out, "{}", suggestion.dimmed()).ok();
     }
-}
-
-// ── Drives ─────────────────────────────────────────────────────────────
-
-/// Render the drives list output.
-#[must_use]
-pub fn render_drives_list(data: &DrivesListOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_drives_list_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data)
-                .unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_drives_list_interactive(data: &DrivesListOutput) -> String {
-    let mut out = String::new();
-
-    if data.drives.is_empty() {
-        writeln!(out, "{}", "No drives configured.".dimmed()).ok();
-        return out;
-    }
-
-    // Pre-compute status strings (avoids formatting twice per entry).
-    let status_strs: Vec<String> = data
-        .drives
-        .iter()
-        .map(|d| format_drive_status(&d.status))
-        .collect();
-
-    let label_w = data
-        .drives
-        .iter()
-        .map(|d| d.label.len())
-        .max()
-        .unwrap_or(5)
-        .max(5);
-    let status_w = status_strs.iter().map(|s| s.len()).max().unwrap_or(9).max(9);
-
-    // Header.
-    writeln!(
-        out,
-        "{:<label_w$}   {:<status_w$}   {:<10}   {:>8}   ROLE",
-        "DRIVE", "STATUS", "TOKEN", "FREE",
-    )
-    .ok();
-
-    for (entry, status_str) in data.drives.iter().zip(&status_strs) {
-        let status_colored = color_drive_status(&entry.status, status_str);
-        let token_str = format_token_state(&entry.token_state);
-        let token_colored = color_token_state(&entry.token_state, &token_str);
-        let free_str = match entry.free_space {
-            Some(b) => format!("{b}"),
-            None => "\u{2014}".to_string(),
-        };
-        let role_str = entry.role.to_string();
-
-        writeln!(
-            out,
-            "{:<label_w$}   {:<status_w$}   {:<10}   {:>8}   {}",
-            entry.label, status_colored, token_colored, free_str, role_str,
-        )
-        .ok();
-    }
-
-    out
-}
-
-fn format_drive_status(status: &DriveStatus) -> String {
-    match status {
-        DriveStatus::Connected => "connected".to_string(),
-        DriveStatus::UuidMismatch => "uuid mismatch".to_string(),
-        DriveStatus::UuidCheckFailed => "uuid unverified".to_string(),
-        DriveStatus::Absent { last_seen } => {
-            if let Some(ts) = last_seen {
-                if let Some(duration) = format_absent_duration(ts) {
-                    format!("absent {duration}")
-                } else {
-                    "absent".to_string()
-                }
-            } else {
-                "absent".to_string()
-            }
-        }
-    }
-}
-
-fn color_drive_status(status: &DriveStatus, text: &str) -> String {
-    match status {
-        DriveStatus::Connected => text.green().to_string(),
-        DriveStatus::UuidMismatch => text.red().to_string(),
-        DriveStatus::UuidCheckFailed => text.yellow().to_string(),
-        DriveStatus::Absent { .. } => text.dimmed().to_string(),
-    }
-}
-
-fn format_token_state(state: &TokenState) -> String {
-    match state {
-        TokenState::Verified => "ok".to_string(),
-        TokenState::New => "new".to_string(),
-        TokenState::Mismatch => "MISMATCH".to_string(),
-        TokenState::ExpectedButMissing => "MISSING".to_string(),
-        TokenState::Recorded => "recorded".to_string(),
-        TokenState::Unknown => "-".to_string(),
-    }
-}
-
-fn color_token_state(state: &TokenState, text: &str) -> String {
-    match state {
-        TokenState::Verified => text.green().to_string(),
-        TokenState::New => text.yellow().to_string(),
-        TokenState::Mismatch | TokenState::ExpectedButMissing => text.red().to_string(),
-        TokenState::Recorded | TokenState::Unknown => text.dimmed().to_string(),
-    }
-}
-
-/// Format an ISO timestamp as a human-readable absent duration from now.
-/// Reuses `format_duration_short` from plan.rs for consistent formatting.
-fn format_absent_duration(timestamp: &str) -> Option<String> {
-    let ts = chrono::NaiveDateTime::parse_from_str(timestamp, "%Y-%m-%dT%H:%M:%S").ok()?;
-    let mins = chrono::Local::now()
-        .naive_local()
-        .signed_duration_since(ts)
-        .num_minutes();
-    if mins < 1 {
-        None
-    } else {
-        Some(format_duration_short(mins))
-    }
-}
-
-/// Render the drives adopt output.
-#[must_use]
-pub fn render_drives_adopt(data: &DriveAdoptOutput, mode: OutputMode) -> String {
-    match mode {
-        OutputMode::Interactive => render_drives_adopt_interactive(data),
-        OutputMode::Daemon => {
-            serde_json::to_string_pretty(data)
-                .unwrap_or_else(|e| format!("{{\"error\": \"{e}\"}}"))
-        }
-    }
-}
-
-fn render_drives_adopt_interactive(data: &DriveAdoptOutput) -> String {
-    let mut out = String::new();
-    match &data.action {
-        AdoptAction::AdoptedExisting { .. } => {
-            writeln!(
-                out,
-                "Adopted {} \u{2014} existing token accepted, sends enabled.",
-                data.label.bold()
-            )
-            .ok();
-        }
-        AdoptAction::GeneratedNew { .. } => {
-            writeln!(
-                out,
-                "Adopted {} \u{2014} new token generated, sends enabled.",
-                data.label.bold()
-            )
-            .ok();
-        }
-        AdoptAction::AlreadyCurrent => {
-            writeln!(
-                out,
-                "{} already adopted \u{2014} token is current.",
-                data.label.bold()
-            )
-            .ok();
-        }
-    }
-    out
-}
-
-// ── Guided error messages ───────────────────────────────────────────────
-
-/// Format a subvolume chooser message for commands that require a subvolume argument.
-/// Names are sorted alphabetically for easy scanning.
-#[must_use]
-pub fn format_subvolume_chooser(command: &str, names: &[&str]) -> String {
-    let mut out = format!(
-        "Usage: {command} <subvolume> or {command} --all\n\nAvailable subvolumes:\n"
-    );
-    let mut sorted = names.to_vec();
-    sorted.sort();
-    for name in &sorted {
-        writeln!(out, "  {name}").ok();
-    }
-    out
 }
 
 // ── Tests ───────────────────────────────────────────────────────────────
@@ -6764,21 +4802,8 @@ mod tests {
         );
     }
 
-    #[test]
-    fn backup_summary_suppresses_unchanged() {
-        let skipped = vec![SkippedSubvolume {
-            name: "sv1".to_string(),
-            reason: "unchanged \u{2014} no changes since last snapshot (21h ago)".to_string(),
-            category: SkipCategory::Unchanged,
-        }];
-        let mut out = String::new();
-        render_skipped_block(&skipped, &mut out);
-        // Unchanged is positive info — should be suppressed in backup summary
-        assert!(
-            !out.contains("unchanged"),
-            "backup summary should suppress unchanged skips, got: {out}"
-        );
-    }
+    // (backup_summary_suppresses_unchanged moved with render_skipped_block
+    //  into voice/backup.rs's #[cfg(test)] mod tests.)
 
     // ── Emergency rendering tests ──────────────────────────────────────
 
@@ -6892,28 +4917,6 @@ mod tests {
             output.contains("Still below threshold"),
             "should show still critical: {output}"
         );
-    }
-
-    // ── Subvolume chooser tests ────────────────────────────────────────
-
-    #[test]
-    fn subvolume_chooser_contains_usage_and_names() {
-        let output = format_subvolume_chooser("urd retention-preview", &["docs", "pics", "home"]);
-        assert!(output.contains("Usage: urd retention-preview <subvolume>"));
-        assert!(output.contains("--all"));
-        assert!(output.contains("Available subvolumes:"));
-        // Should be sorted alphabetically
-        let docs_pos = output.find("docs").unwrap();
-        let home_pos = output.find("home").unwrap();
-        let pics_pos = output.find("pics").unwrap();
-        assert!(docs_pos < home_pos && home_pos < pics_pos, "names should be sorted");
-    }
-
-    #[test]
-    fn subvolume_chooser_single_name() {
-        let output = format_subvolume_chooser("urd history", &["only-one"]);
-        assert!(output.contains("only-one"));
-        assert!(output.contains("Usage: urd history"));
     }
 
     // ── humanize_duration tests ────────────────────────────────────────
